feat(proxy,cli): the defect counters get somewhere to be read
#818 surfaced per_model_defects on /v1/proxy/status and stopped at the wire.
Neither downstream mirror was updated, so every counter added across the
pivot — stream errors, truncations, empty and reasoning-only turns, dialect
residue, unvalidatable schemas, normalization errors — was readable only by
curl | jq. ADR 0006 kept these instruments on the grounds that a person reads
them; there was nowhere to.

`gglib proxy dashboard` now renders them, in the surface that already streams
this endpoint. Two shapes of the data drive the layout:

Only models with something to report get a line. A healthy model produces zero
of every counter, so listing the clean ones would bury the one that is not —
and an empty section is a real answer rather than a missing one. "Nothing
recorded yet" and "none across N requests" are kept distinct, because before
anything is forwarded there is no evidence either way.

reasoning_only is shown inside the empty total, not beside it, since it is
counted within empty_responses; peers would read as more empty turns than
happened. Repairs read as a ratio, which is the number worth watching.

The file was already the largest in the crate, so this splits it on the line
that makes the interesting part testable: wire.rs (the mirror), render.rs
(pure, snapshot to text), render_tests.rs, and mod.rs for the IO. The ratchet
baseline records the result — one 1255-line file became four, largest 624.
